Setting major update to the app using the latest modification of the logic of the application

+++ applications/authentication/models.py
@@ -14,7 +14,7 @@ class CustomUser(AbstractBaseUser, PermissionsMixin):
     profile_img = models.ImageField(upload_to=upload_to, blank=True)
     birthdate = models.DateField(null=True)
     card_number = models.CharField(max_length=15, null=True, blank=True, unique=True)
-    solde = models.IntegerField(default=5000, blank=True)
+    solde = models.DecimalField(default=5000.00, max_digits=15, decimal_places=2, blank=True)
     # is_admin = models.BooleanField(default=False)
     date_joined = models.DateTimeField(auto_now_add=True, editable=True)
     is_active = models.BooleanField(default=True)
